When applying integral conversions, ignore qualifications
There is a check when building a standard conversion sequence to avoid applying
integral conversions when the type is the same, but it does not drop the
qualifications, and a case like 'const int' and 'int' look like different,
causing an integral conversion to be performed. This causes conversions
like 'const int& -> int' to be worse than they should.