Make UDRInfoItem work on a Symbol
Since OpenMP core is able to solve the exact symbol (except
for dependent contexts where nothing happens) for a UDR,
use it since eases lots of things.

Anyway, it is still possible to build UDRInfoItems based
only on the name. C/C++ OpenMP reductions already defined
in OpenMP 2.5 are signed in this way.