Make PragmaCustomConstruct::get_declaration kinder
If the user calls get_declaration it is assumed that a declaration
will follow the pragma. It turns that inside a function, there
are no declarations "per se" but declaration-statements (which
simply hold a declaration...) so make Declaration kinder
when wrapping statements that are just declaration statements
and also make PragmaCustomConstruct::get_declaration try to
get a declaration off a declaration statement.

This commit fixes #109