Parse config files using a full fledged parser
This commit removes mcfg.{c,h} which was a hand-made parser for configuration
files. Due to dubious licensing issues (the file was not proper licensed and
we believe it was based on some code from Samba) we decided to replace it
with a typical parser built using flex and bison.

Since mcfg.{c,h} are gone, the library libmcxx-mcfg.so is not needed
anymore, so should you find in your lib installation, due to a previous
install, you can delete it.

Lexer is in src/driver/cxx-configfile-lexer.l and uses lots of
exclusive states from Flex, which makes it slightly more involved
than expected.

Parser is in src/driver/cxx-configfile-parser.y and replaces several
functions present in cxx-driver.c

This new parser uses some features of bison not used in the main
C/C++ frontends which I believe they will prove useful and
will simplify some aspects of them.

This commit fixes bug #139