Changes to improve tasking support
Currently you can use an arbitrary expression. If the expression itself
is an id-expression (so it names a symbol) the dependency will be built
upon the outline argument, otherwise an additional pointer will be added
in the outline argument (this one will point to the expression of
the dependency) and the dependency built upon this pointer.

Hopefully this commit fixes #155 and fixes #158