There was some laxity in the spu builtins code related to integer types
For sure this will break some codes because of some "type sloppiness" when
calling the builtins. Should this happen a cast in the arguments should be
enough to fix these problems.

The reason behind this behaviour was originated by Altivec not
precisely defining the compatibility among vector type (vector of char
vs vector of integers).

This commit fixes ticket #264