Fix for member declarations which declare types
template<int _N>
struct A
{
    union u { char c[_N]; };
    u data_;
};

the type of data_ is properly set to [A<N>]::u
while in the case

template<int _N>
struct A
{
    union u { char c[_N]; } data_;
};

data_ type was set to the type of union u. While this seems
the obvious thing to do, instantiation may need to update
data_ type and this is only possible if the type is set
to a dependent typename like [A<N>]::u. For instance in

  A<3> a;

instantiation needs to finish the class (adding implicit
constructors and such). data_ member is found to be a
dependent type class (actually A<N>::u) which has not
been finished thus it lacks constructs, etc.

Make member declarations which declare types inside dependent
classes to be dependent typenames (like it happens with declarations
which did not, e.g: using a simple type specifier of a member typename
of the dependent class)

Also refactor some enum related code just before the change of the type.

New test added.