Set the underlying type earlier than the declarators
In a declaration like

enum E
{
    A,
    B = A + 3
}

the underlying type of the enum type is not set until all enumerators have been
signed in. When checking expression 'A + 3' we promote the enumerator type from
enum to its underlying type but that has not been set yet. And then the typechecking
fails miserably.

This early setting of the underlying type clashes with the fact that the enum
should have a type that accomodates the real range of the enumerators values,
in particular in C++. I guess the right approach is for each enumerator store
its appropriate integer type and for the whole enum the widest type of each of
the enumerators.

See #345