Instrumentation of user functions in gpu.
This commit is related to #454. As GPU kernel calls are asynchronous,
we cannot insert the closing events of the outline at its end, because
the kernel might be still running. We have to insert them right after
waiting for the kernel to finish, which is done inside the Nanos++
runtime.