Add support for conditional and '<' operator in hlt simd
The operator '?' is implemented using bithacks and it's not pure.
The operator '<' is implemented using a gcc builtin only available in Intel Architectures.
The expansion system of scalar constants to vectors has been changed and now the '__builtin_vector_expansion' is used again.
Expansions are performed on simple constants expressions now.
For example, the expression 1 + 2 + 3 will generate three vectors instead of just one with the scalar operation 1+2+3 replicated. In addition, a naive optimization of constant expressions is applied before the expansion.

It fixes #577