Option to define initialization variables as static
This is required by icc as it cannot define a weak symbol in a specified
section (the symbol is not defined as weak anymore and linking fails).

A new --variable=weaks_as_statics tweaks this, defining the variable as simply
static. This has the drawback that the runtime may receive more than once the
same pointer, although the runtime supports this (only calling each function
once)

This commits fixes #574