Fix ambiguity uncovered while fixing ticket #576
   [1][2] a

could be parsed as (parentheses added for clarity)

   [1]([2] a)     or    [1][2](a)

this did not happen because we only allowed pointers types and a shaping
expression returns an array type so the checking routines automatically solved the
syntactic ambiguity.

Since now we allow an array to pointer conversion in a shaped expression, both
interpretations are OK causing the ambiguity.

Rewrite the grammar in a way that we do not allow a shaping expression be shaping
another shaping expression (unless we use parentheses) so

  [1][2] a

is always interpreted as (parentheses added for clarity)

  [1][2](a)