'smp' SIMD-related pre_run code has been moved to HLT pre_run
The declaration of 'smp' device specific functions have been moved to the HLT phase pre_run.
This means that SIMD stuff will be enable when --hlt flags is on, even when hlt simd is not present.

A new compilation phase should be created and also a specific --simd flag (that also activates --hlt)
to enable SIMD only when necessary.