Add a default function for 'int abs(int)' function
There exists 3 specific SSSE3 intrinsics for char/short/int abs operation.
However, char and short are currectly not supported because there is only one 'int abs(int)' function in stdlib.h
and we cannot add 3 specific functions with the same priority (DEFAULT) to the same generic one and select
the most appropiate according to the type of the parameter.

Also it fixes a bug related with generic functions and reimplements some aspects in the HLD SIMD prettyprint_callback.

It fixes #596