Allow braced-init-lists in return expressions
This is a C++1x feature but g++ is supporting it since 4.4 and it
appears in at least one header since 4.5 (even if -std=c++0x is not
actually enabled)

Currently the compiler was not checking very much the return statement,
now it does.

In a future patch I will port this to nodecl branch where we
need to know which conversors may end being called.

This commit fixes #656

This commit fixes #486