Add --debug-flags=binary_check
Use this flag to verify that the codegen output has the same binary
stuff that would have if we only had invoked the native compiler.

Of course, this is only meaningful with profiles like plaincc, plaincxx
or plainfc. Any profile that modifies in any way the code is expected to
fail.