Remove scope link
scope_link_t (and its C++ incarnation TL::ScopeLink) are gone. This
structure was used to retrieve the context of a given node. Not
something a compiler usually needs, but Mercurium does to allow
subparsing. Given an AST and a scope_link_t you could retrieve
a decl_context_t

Instead of keeping a scope pointer in every tree (which for lots of them
the context seldomly changes and the information would be repeated so
many times) we have created a new nodecl_t node NODECL_CONTEXT which
will hold the context of the hierarchy nested under it.

We introduce a NODECL_CONTEXT in places where the inner trees inherit a
different context. This happens in NODECL_FUNCTION_CODE,
NODECL_COMPOUND_STATEMENT and NODECL_COMPOUND_EXPRESSION. The usual
child of a NODECL_CONTEXT is a sequence of statements (in C/C++ this
list is a singleton of a NODECL_COMPOUND_STATEMENT)

A NODECL_CONTEXT can also have an expression as a child. In many
places in the compiler an expression must be kept along its context
(for instance the size of an array type). If we set the expression
as a child of its context there is no need to carry both things
all the time. This makes expressions standalone when needed.
This not yet in use, it is not a priority now.

This approach using a NODECL_CONTEXT would allow writing phases where
new contexts are created at the developer will and then codegen could
make the most of them (for instance: in Fortran we could avoid creating
unique names and let codegen do that for us). This is not in use yet.

This commit fixes #7