Always register parameters
In a function definition, parameters used in the declarator were kept in
entity_specs.related_symbols. This did not happen for function
declarations, so we keep these (if named). This way we always have
the latest set of declared parameters.

This is very useful for pragma
custom declarations since we can recover a prototype scope level
context. We keep this context (if available) as the fourth son
of a NODECL_PRAGMA_CUSTOM_DECLARATION