Codegen phases are now pluggable
Rewrite in C++ of C++ codegen is not yet complete.

Fortran codegen is not yet ported to C++ but we have a codegen phase
which calls the existing code in frontend.

Open remains the question how to deal with the code in frontend, as we
need it for diagnostics (c_cxx_codegen_to_str and
fortran_codegen_to_str). Maybe the compilation process should offer
a single codegen_to_str and be able to relay it to the base codegen
of the current language.

plainfc/plainf95 profiles have been defined to use legacy fortran
codegen phase

configuration file allows a new option named 'codegen_phase'. It cannot
be repeated.