Fix a problem with intrinsics
In a implicit context, if the name of a variable matches with an
intrinsic name, it won't be always an intrinsic function (can be an
implicit subroutine).

Example:

PROGRAM P
CALL VERIFY(...)
END PROGRAM P

SUBROUTINE VERIFY(...)
...
END SUBROUTINE VERIFY

the program is calling to verify subroutine (not the intrinsic).