Optimize load times of Fortran modules
Loading a big module took about 2.90 seconds. Profiling with sysprof
revealed that load_symbol whas taking about 90% of all the time involved
loading it.

A first observation was that preparing the statement every time was
inefficient (sqlite3_prepare) as it took almost the same time as
executing it (sqlite3_step). So we rewrote load_symbol to use a prepared
statement instead. This gave a modest improvement, the module was then
loaded in 2.20 seconds (1.3X speedup).

Further analysis shown that lot of time was spent doing a SELECT on the
attribute table. Using EXPLAIN QUERY PLAN from sqlite3 we discovered
that a SIMPLE SCAN was being done on a big table, leading to ludicrous
performance during the query. Adding an INDEX for the two used fields in
the SELECT statement improved the time. This improvement was dramatic:
the module can be loaded in just 1.27 seconds (2.25X speedup).

Further improvements are possible, but these relate to module creation
rather than loading.