Solved problem with sql insert statement
The symbol type only can be created when it has been completely
filled. The creation of a type can be recursive (we are creating
a type and before assign it an address we call the create function again).

For this reason, we protect the insert statement: we don't insert if the
same tuple already exist.

It fixes #823