Fix analysis for multiple func calls within expr
It fixes the analysis when one expression has multiple calls to the same function.
Case 1:  Method A has an expression such as  "a + B() + c + B()"
         Call graph will be:
             A()  -->   B()
                  -->   B()
Case 2:  Method A has a call to method B and method B has a recursive call
         Call graph will be:
             A()  -->   B()   -->   B()

IPA analysis is now fixed to do not recognize Case A as a recursive call.