Add Liveness computation for complex epressions
Compares not only whole nodecls but parts of them. E.g.:
    struct STR {
        int a;
    };
    void foo (STR* str)
    {
        STR* str_2 = str;
        int b = str->a;
    }
Inside 'foo' we make two uses: 'str' and 'str->a'.
Since the 'str->a' is a part of 'str' we will only show the use of 'str'

We still have to implement other cases like a part of an array has
an undefined behaviour and then all the array is used, we have to include
in the use list only the part that had not an undefined behaviour.