Derived type specifiers can have the name of a Fortran keyword
MEMORY is a Fortran 2008 keyword (related to coarrays, if I
recall correctly)

I'm not sure why I explicitly forbade using Fortran keywords in
derived_type_specifiers, but it does not make any sense to me.
A derived type name is used in a way that it cannot be confused as
a non-expression Fortran statement (which must start by Fortran keywords).

This commit fixes #916