Do not add modules found in other modules in the cache
The aim of the module cache is to avoid loading module files.

If we find a module symbol inside another module and we cache it, later
requests to load the module will be directly resolved through
the cache. This will be a wrong hit because that module symbol
found may be incomplete.

This commit fixes #935