Fixes to support SECTION as a special directive
OpenMP 'section' syntax does not quite fit within the usual distinction
of "directives" and "constructs".

What we do is registering them as directives. Then, since they only appear
inside 'sections' (or 'parallel sections') we promote them to constructs,
properly grouping the statements that they cover. Once this grouping has
happened a 'sections' (or 'parallel sections') is not a sequence of
'section' and regular statements but a sequence of 'section'.
Each 'section' has as child the sequence of statements it covers.

Just a little glitch due to C/C++: Such list is not an immediate child
of the 'sections' (or 'parallel sections') pragma but instead of a
compound-statement required by the C/C++ OpenMP syntax. That compound
is the immediate child of the 'sections' (or 'parallel sections').

This commit fixes #759