Mitigate linker issues in newer Linux distributions
Modify linker_options in {{{10.config.omp-base}}} so nanos libraries are not
handled {{{--as-needed}}}, since some symbols may not be explicitly used by the
program but Nanos++ runtime.

This is a long standing issue for Nanos++. I've tested it in a Ubuntu and it seems
to work, both for OpenMP model and OmpSs model (when ssCompatibility != NULL)

Longer explanation follows:

In some distributions (like Ubuntu) {{{--as-needed}}} linker flag is enabled by
default. For some technical reasons beyond the scope of this commit, the final
executable linked must depend on all Nanos++. Here "depends" means the ELF file
should have DT_NEEDED entries for {{{libnanox.so}}}, {{{libnanox-c.so}}} and
{{{libnanox-{ss,omp}.so}}}.

The linker, running under {{{--as-needed}}}, does not detect any dependence
between the executable program and {{{libnanox.so}}} (but it '''DOES''' for
{{{libnanox-c.so}}} and {{{libnanox-{ss,omp}.so}}}), thus the executable
program does not have have a {{{DT_NEEDED}}} entry for {{{libnanox.so}}}. This
causes a myriad of subtle problems and hard to debug errors in applications,
even in their startup, when Nanos++ is initializing.

See Nanos++ ticket [nanox:#570]

See #752