Rework the way we emit outlines
Instead of passing a Nodecl, we will pass a Source plus a nodecl
with appropiate context. Note that this Source will usually have
a placeholder that we will replace once the outline has been
emitted.

This was necessary for "#pragma omp for" and "!$OMP DO" since they
need to create new code. Appropiate context creation steps are taken
for Fortran.