Several fixes
Replace instantiate_template_class with two functions
instantiate_template_class_if_needed and
instantiate_template_class_if_possible. The former means always
instantiate because we will check inside the class, the latter means
try to instantiate if possible but not an error if we can't.

Add also a is_instantiated attribute to tell if a specialization has
been instantiated already. Set this attribute to true when an instantiation
ends or when we are defining an explicitly specialized type

Fix a wrong prettyprint in codegen involving anonymous unions