Wrong clauses warning is now managed by Core phase
Warnings for wrong clauses used to be managed by nanos phase.
This was useless for function level constructs since the AST
was already modified when the compiler reaches this phase.
Now we have a flag in phase Core, where the trees are still
as they are in the input code, that enables showing warnings
only the first time Core phase is executed

This fixes #475