Fix replacement flow
Due to recent changes to Nodecl::NodeclBase::replace it is not possible
to replace twice a node if the tree is a placeholder.

Until now some source was parsed when the body was outline. Just pass a
simple nodecl to act as a placeholder and parse whatever is needed using
that placeholder.