Properly generate the argument of nanox_get_ptr
Currently we are not storing lower bounds for assumed shapes like

 FUNCTION F1(X)
  INTEGER :: X(:)
  ...
 END

It defaults to 1, and only array POINTERs could legitimally have a
NULL lower bound. We should fix this. But this will correctly default
to 1 as expected, though this must be fixed it should not cause
any problem now.

For this case, though

 FUNCTION F2(X)
  INTEGER :: X(5:)
  ...
 END

we have to make sure our get_ptr function is

  FUNCTION NANOX_GET_PTR_xxx(T)
    INTEGER, TARGET :: T(5:) !! Note the "5:", ":" would be wrong
    INTEGER, POINTER :: NANOX_GET_PTR_xxx(:)

    NANOX_GET_PTR_xxx => T
  END

Otherwise our pointer would have a wrong base and indexes based on T
would not work on the returned pointer .I.e. NANOX_GET_PTR_xxx(5:10)
would be T(10:15)!

Currently the only case that does not work is when lower bound is
nonconstant. We may have to extend NANOX_GET_PTR_xxx with extra
parameters in this case (just the array would not do)