Register generic names first when encountered
Until now we registered a generic name after all the interfaces.
Then we added every interface in the set of related entries of the
generic name.

The problem with this approach occurs when codegen is emitting a module.

If the (non-module procedure) interfaces are found first then their
(non generic interfaces) are emitted first. Later the generic interface
is emitted accordingly but now it is too late, as codegen realizes
does not emit the specific interfaces of the generic one (as they were
already emitted since they were found first!)

By first registering the generic name, codegen encounters it first
in the list of members of the module, thus it is emitted first, along
its components. Because of the order, the components have not been emitted
yet, so they appear declared inside the generic interface, as we meant.

Technically a specific interface can appear in more than one generic
interface, but such thing cannot have in a single program unit, so this
should not be a problem right now.

See #1052