Early exit when a comma expression has one of its arguments wrong
 TYPENAME * f, g();

is first interpreted as

 (TYPENAME * f), g()

the rhs of the comma operator will fail but nevertheless we check
g() as well, registering a fake g() causing a wrong type be emitted.

This is also caused by the fact that g() does not have prototype and
we are laxer than usual updating types of functions.