Fixes during determination of user defined conversion sequence
When determining a user defined conversion sequence by means of
conversor constructors, it may happen that two conversor constructors
get chosen because of valid standard conversions to their first argument.

The standard is unclear at this point.

But we understand that, once we have a set of viable constructors that
might be called, we have to choose one which is the best. We first thought
on reusing solve_constructor but this comes at risk of an unwanted
recursion (thought this could be cut off by means of more flags in a
hypothetical solve_overload_internal function).

So what we do is replicate part of the effort performed by solve
constructor  but limiting ourselves to standard conversions sequences to
the first argument.