Add a new field 'friend_candidates' to entity specs
Every template-id friend function declaration in a
template class (that is, in a dependent context) will
have a set of 'friend_candidates'.

A template-id friend function declaration refers to
an pecialization of a an existing function template. We will
store in this set the possible template specialized functions.

Example of use:

template < typename _T, typename _S>
void f(_T*, _S);

template < typename _T>
struct A
{
    friend void f<>(_T*, _T*);
   // The candidate friends fo this template-id friend
   // declaration will be { ::f<T,S>(_T*,S) }
};