Make sure we do not mix intrinsic functions
Overhaul all the code related to intrinsic resolution. Up to now
the code could end mixing two different interfaces. Avoid this
by clearly splitting different interpretations of intrinsics.

The problem in #1063 not only affected SUM but functions behaving
similar like IALL, IANY, IPARITY and PRODUCT.

This rework also fixes a problem with ALLOCATED. The code was
unable to tell apart ALLOCATED(SCALAR) from ALLOCATED(ARRAY).

Note that the type of the intrinsics is still represented as computed
function types but the compute function is not meant to be executed,
instead the code should use the Fortran own routines.

This commit fixes #1063