Add indexed attributes is_parameter and parameter position
In ticket #834 we removed the attribute parameter_position since it is
wrong when a symbol may be a parameter of more than one function (as it
happens in Fortran for statement function statement and ENTRY)

On the other hand, C/C++ allow declarations be repeated (though
no symbol appears in more than one function as a parameter). Having
removed the parameter position attribute it is absolutely impossible
to tell the position of a parameter in some context. In particular
when parsing #pragma at the declaration level (see #1064) may require
issuing a diagnostic for unnamed parameters.

So, is_parameter attribute has been removed. Instead a map called
function_parameter_info is stored. Such map goes from a scope_entry_t*
(a function) and will return us the information of this symbol as a
parameter of that function. In the case of Fortran this means that a
symbol may have more than one entry with different parameter positions.
In the case of C/C++, only one parameter position will be kept.

Now querying whether a symbol is a parameter requires the function in
which we want to see if it is a parameter. We can also check if the
symbol is a parameter of any function in context by just checking that
num_function_parameter_info != 0. This may be required in Fortran.