Do not insert an intrinsic name if called in an initialization
Since the ambiguity code relied on this being done all the time
we stop ambiguity checking if one is found to be OK (this makes
things slightly faster as a side-effect).

Now we are not able to detect two valid interpretations. Hopefully
an error like this will be uncovered in other checks.