Adapt analysis singleton to many phases uses
Analysis Singleton is the object that allows the execution of any analysis
and Memento is the object that contains the states of the analysis.
This way, any phase and different stages in the same can call the same analysis
for different modifications of the same input code. Users only need to create
a new Memento and use the Analysis Singleton to call the analysis.

See [e9a5cedfa86460044e9cdad28707f99df0738732]