Add MERCURIUM_LOC intrinsic
This intrinsic returns a "void*". A call like MERCURIUM_LOC(x) is
simplified to &x which in Fortran codegen means LOC(x).

This intrinsic is not meant to be used by Fortran users but it is an
internal intrinsic used during transformations.

Remember that void* in Fortran is emitted as INTEGER(8) (or 4 in 32-bit).