Lots of fixes in OutlineDataItem and OutlineInfo
An OutlineDataItem must have a symbol now, thus most checks of the
validity of the symbol should be considered redundant.

Add a mechanism to automatically synthesize the boundaries of arrays
passing them as firstprivates. Remove ItemKind as it is redundant now.

Rename _shared_expression to _base_expression, representing the argument
of the call in a function task.