Improve the algorithm of class lookup
Profiling with sysprof showed that this function took a lot of time
memsetting a large array used later for results in base classes.

Rewrite the algorithm so it checks the bases if actually needed.
Now it only allocates for the worst-case where all
nondependent base classes contribute a result in the query.