Minor fixes in analysis
- Allow non l-value nodecls as extended symbols for nodecl comparison.

- Modify the extensible's graph more external graph_node. Now it has not
EXTENSIBLE_GRAPH type but a type depending on the nodecl it contains.
This is necessary when creating graphs from nodes which are not
function code; otherwise, the results are not correct.

- Use INT_MAX from <climits> instead of a very big integer to define
an arbitrary identifier for the graph nodes. Same for the level of
nesting desired for analysis in the vectorizer.