Fix interface for nodecl matching in Ind. Vars. analysis
Most external graph can contain the same nodecl than an inner
node when the PCFG is built from a non TopLevel nodecl.
For instance, building a graph from a ForStatement 'n' will
result in an EXTENSIBLE_GRAPH node containing 'n' and a
LOOP_NODE containing 'n', as well.
When trying to find the PCFG node corresponding to 'n', the
analysis must return the LOOP_NODE, which contains the
loop analysis, instead of the EXTENSIBLE_GRAPH.