Do not check this unless really needed
When figuring if a call to a function has dependent arguments, we were
too early checking if this was dependent. There are a number of constraints
to be fulfilled before such check is valid.

When the check was done too early it also was done for cases where it
should not have been done. This caused that some calls were too lightly
considered as dependent, causing unpredictable results.

This commit fixes #1291