Recognize common C library methods in Use-Def analysis
Functions that code is not available in the translation unit
cause Use-Def analysis to determine undefined the usage of
global variables and reference parameters.
Common C library calls are tagged, so now the compiler is able
to know that these functions do not modify certain variables and
the analysis is less limited.

See #859.