Fixes when synthesizing the type of an array in the argument structure
We realize too late that the field of the argument structure will be
an ALLOCATABLE array, requiring a descriptor. So we, wrongly, create
an array without descriptor to an array with descriptor. This is not valid
in the internal representation, so the compiler ignores that dimension
when emitting the code.

This commit fixes #1304