Minor fixes in Use-Def analysis
Usage propagation between nodes and its outer nodes was not done
in the proper order (first propagate kill and undef sets, and then
upper exposed, otherwise killed or undef variables may be classified
as upper exposed as well, and this is not correct).

Usage in recursive calls was not implemented. We add minor implementation
that only parses value passed parameters. We have to propagate the
analysis of reference parameters and global variables as well.