More fixes
VLAs require a bit more of work.

When the outline function calls the unpacked function, we add a cast to
make sure the type of the structure is handled accordingly. Note that
for VLAs the structure of arguments just keeps a void pointer.

Reminder: in the argument structure we keep two kinds of entities,
captured entities keep a value (note that the value might be a pointer
if we are capturing a pointer) while shared entities keep a pointer to
the data. In order to avoid copying twice the data in the call to the
unpacked function (from the outline function) the unpacked function receives
references for these entities. This applies both for shared and captured
entities. A reference to a shared entity will be bound to the address
of the entity. A reference to a firstprivate will be bound to the field
of the structure itself.

Remember that while arrays are entities on its own in C/C++ (and this
includes VLAs in C) they cannot be passed by value. So the only thing we
can pass actually is a pointer to somewhere we know our array starts In
C, firstprivate arrays that are VLA are allocated in a special way at
the end of the structure, and a void pointer to that location is set up.
Shared arrays only require a pointer to an array with the caveat that VLA arrays
will use a void pointer too.

So when passing an entity that is a VLA (or derived from VLA like a
pointer to an VLA) we need to ensure that its type, which will always be
a void pointer as explained in the previous paragraph, is properly
handled to the unpacked function. i

For shared VLAs, the void pointer is just casted to an appropiate
pointer to the array and then derreferenced. This yields a reference to
the array, the type exactly expected by the unpacked function. You won't
see such dereference, as codegen will omit it when it discovers that you
are dereferencing an argument passed to a reference type parameter.

For firstprivate VLAs, the void pointer is the actual storage of the
pointer itself. So we have to convert it into a reference of the type of
the parameter of the unpacked function. Note that no dereference is
required in this case, as the lvalue of the field will be used to bind
the parameter reference.