Add Auto-scoping as a phase
This is a first approach tested with few examples such as Fibonacci.
The phase is added after OpenMP parsing, and uses the clause default(AUTO)
attached to a Task to launch the Automatic Scoping process for this task.
In this commit, we do not yet modify the internal AST, so the compiler
still returns an error because the scope of the variables has not been
automatically decided.