Enhance Induction Variables Analysis
We recognize now variables that are constant in a loop. So,
variables that, within a loop, are always increased or
always decreased by a variable that is constant within the loop,
are now recognized as an induction variables.

This commit also adds a missing visitor in the Use-Def analysis which
caused the lhs of an AddAssignement not to be classified as Killed.