Added an API to detect unused clauses
Also use it for tl-omp-base for emitting diagnostics.

The API allows to mark and unmark clauses as used or unused. Currently
it does not allow detection of an unused "parameter" clause (i.e. the clause
that follows right after the pragma construct, like in
"omp critical"). But these are less unlikely to be wrongly used.

This process is not as automatic as I'd like it to be but a bit of
discipline is enough to make it work.

This commit fixes #1301