Make Nodecl::RangeLoopControl more analysis amenable
The induction variable was not explicitly kept in the RangeLoopControl
as a symbol occurrence. This would force analysis routines which
look for occurrences to deal this node in a special way.

Remove the symbol of the induction variable from the node and instead
add a child with an artificial occurrence.

This commit fixes #1402