Add support for Task multiple post-synchronizations
Support for tasks that can be synchronized in different points.
Also adds a new node representing a synchronization that occurs
after the finalization of the function that creates a given task.
There is a synchronization edge between the given task and this
new node when we cannot ensure the synchronization occuring
within the function ( synchronizations in conditional branches
or no synchronization at all ).

Example:
    void foo(int y)
    {
        int x;

        #pragma omp task out(x)
        {
            x = x + 1;
        }

        if (y > 1)
        {
            #pragma omp taskwait
        }
    }

Fixes #1399.