All the return arguments should be added as alloca input depedences
This is a very important thing in nested function calls:
{
	x = foo(foo(y));
}

Our visitor transforms this code into:

{
	int *tmp1 = NULL, *tmp2 = NULL;
	foo__(y, tmp1);
	foo__(tmp1,tmp2);
	*x = *tmp2;
}

The expression *x = *tmp2 will be taskified and Its work descriptor should have space for the two temporal variables. This is because the second call to foo does not allocate space for the return argument of the first call.