Support for GCC C nested functions
In order to support GCC C nested functions and make NODECL_FUNCTION_CODE
a simpler and more consistent node accross the three languages (only
two of them actually supporting nested functions), the third child
containing a list of NODECL_FUNCTION_CODEs has been removed.

That child was used in Fortran to store the internal subprograms.  Since
in C, nested functions can be declared anywhere in block scope, we
believed that allowing nested functions in the body of the function and
having a differentiated node for the Fortran internal subprograms would
end in a redundant representation.

Thus, in Fortran, internal subprograms will be represented as normal
nested functions (though remember that in Fortran only one level of
nesting is actually allowed while such restriction does not exist in GNU
C) and as such they will appear in the body of the function.

Note that the position inside the body of the function is only relevant
for GNU C which follows the usual "declared earlier" approach. In
Fortran the position is not relevant, since the enclosing procedure sees
all the internal subprograms and all the internal subprograms see their
siblings and its enclosing program unit. So, in Fortran, the position
of the nested functions should be regarded as nonrelevant (although
fortran03-buildscope.c will always put them at the end but do not rely
on this when doing analyizing/transforming).

This commit fixes #1471