Add two strategies to mitigate unused variables warnings
A first strategy simply adds __attribute__((unused)). This strategy
is the safest except that it requires gcc compatible attributes. This
strategy is, though, not the default.

You can enable this strategy using the flag:

  --variable=emit_saved_variables_as_unused:1

The default strategy removes all the unused saved expressions that do
not appear in the body of the code. This strategy is riskier but we
believe it is a better approachi so we made it the default.  The removal
happens in C/C++ codegen just before emitting a NODECL_FUNCTION_CODE

You can disable this strategy using the flag :

  --variable=prune_saved_variables:0

This commit fixes #1461