The 'priority' and the 'if' clause should not generate taskwaits
Even if their arguments are parameters with an input value dependence.
This means that the value of the priority (and the expression of the if clause)
will be captured during the task creation without do any previous taskwait.

Partial revert of commit: 2cbe3e47bd1c6a00839875c8d837a9d46ebb73de