Add support for functions with masks
Functions now can be called from regions with masks and 'mask' and 'nomask'
clauses work fine. Only 'nomask' is allowed if the underlying architecture
doesn't support masks. If neither of them are present, the compiler will emit
both versions only if the architecture supports masking.

It should work but it cannot be tested until ticket 1504 is solved.

See #1502