Support +:int SIMD reductions in SSE/MIC
Add support for SIMD reductions in SSE and MIC.
Current supported reduction ops & types: +:int

To use SIMD reductions use the following Vectorizer interface:

- is_supported_reduction: determines if the reduction is supported or not on SIMD instructions.
If it is not supported, a scalar approach should be used.

- vectorize_reduction: vectorizes reduction and returns a list of pre-nodecls and post-nodecls
that should be placed accordingly before and after the reduction loop/parallel.

Please, Roger, add as many tests as possible, combining all the pairs op,type that could potentially
be vetorized in OpenMP. They will be vectorized as long as I add support for them.

Support for 'omp simd for' is still being implemented.

See #1508
See #659