Implement WFV masking algorithm
WFV stands for Whole Function Vectorization.
We extended this algorithm setting output mask to all-zero in a basic block
with a return statement.

Our implementations still needs to tackle what to do with recursive
function calls, where all-zero masks should be detected in order to
stop the execution (and the recursivity).

See #1563