More interesting MaskCheck heuristic
This commit introduces a more sophisticated heuristic.
The outer BB propagates its cost value to the nested one.
Once the threshold is reached, the cost value is reset to zero again.
This way, a BB block could produce a mask check both if it exceeds the
cost threshold by itself or, in case that the previous BB(s) didn't,
the cumulative cost do it.

It also change the masking model. Now, no mask is represented with
a MaskLiteral set to all one.

In addition, a new operation 'const_value_is_minus_one' has been added.

If fixes #1568