Changes in is_induction_variable_dependent_access
The inner method 'is_induction_variable_dependent_access' is now
called 'contains_induction_variable' and it is intended to be more
generic. It could be use not only for memory accesses.
Please, feel free to give a more appropriate name.

AdjacentAccessVisitor now checks if the pointer (subscripted exp)
of an array subscript is an IV, not only the whole array subscript
or its subscripts. Maybe an independent visitor to check if a Nodecl
contains inductions variables could be more appropriate instead of using
AdjacentAccessVisitor to do this.

These changes need to be deeply checked out as currently the analysis still
shows some warnings:

'Induction Variable 'map' not found in the RD_IN set of loop '12''

See #1544