Further information about ticket #1554
Analysis computation is now centralized in the tl-vectorizer.cpp (initialize_analysis).
I've been having a look and I think I'm using the analysis appropriately.
Please, correct me if I'm wrong.

New analysis is computed twice for the fib function: one for the mask version and another
one for the nomask version.
Then the analysis is computed once for the main function when the first loop is going to be vectorize.
For the second loop, the vectorizer detects that the enclosing function of the loop is the same as
the current computed analysis and then it reuses it, as it was planned to prevent the analysis to deal
with VECTOR nodes (of the previous vectorized loop).

However, when the vectorizer tries to vectorize the second loop, nodes seem not to be in the
analysis: "... is not contained in the current analysis...". Vector nodes show up since this is
a warning an the vectorizer continues its task (this is not a problem).

So, Sara, we need your help!

See #1554